task-5: Create lambdas for importing and parsing csv files.
