+++ files/test.csv
@@ -1,0 +1,9 @@
+count,description,id,price,title
+1,Short Product Description1,7567ec4b-b10c-48c5-9345-fc73c48a80aa,2.4,ProductOne
+1,Short Product Description3,7567ec4b-b10c-48c5-9345-fc73c48a80a0,10,ProductNew
+1,Short Product Description2,7567ec4b-b10c-48c5-9345-fc73c48a80a2,23,ProductTop
+1,Short Product Description7,7567ec4b-b10c-48c5-9345-fc73c48a80a1,15,ProductTitle
+1,Short Product Description2,7567ec4b-b10c-48c5-9345-fc73c48a80a3,23,Product
+1,Short Product Description4,7567ec4b-b10c-48c5-9345-fc73348a80a1,15,ProductTest
+1,Short Product Descriptio1,7567ec4b-b10c-48c5-9445-fc73c48a80a2,23,Product2
+1,Short Product Description7,7567ec4b-b10c-45c5-9345-fc73c48a80a1,15,ProductName
